// webapp/src/api/model/armor.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { EArmorSlot } from './eArmorSlot';
import type { ArmorEffectsItem } from './armorEffectsItem';
import type { Material } from './material';
import type { ERarity } from './eRarity';

export interface Armor {
  armor: number;
  armorSlot: EArmorSlot;
  description: string;
  effects: ArmorEffectsItem[];
  id?: string;
  material: Material;
  maximumStackSize: number;
  minimumStackSize: number;
  name: string;
  note: string;
  protection: number;
  rarity: ERarity;
  requirement: string;
  tags: string[];
  tier: number;
  upgradeSlots: number;
  vendorPrice: number;
  weight: number;
}