// webapp/src/api/model/jewellery.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { JewelleryEffectsItem } from './jewelleryEffectsItem';
import type { Material } from './material';
import type { ERarity } from './eRarity';

export interface Jewellery {
  description: string;
  effects: JewelleryEffectsItem[];
  id?: string;
  material: Material;
  maximumStackSize: number;
  minimumStackSize: number;
  name: string;
  note: string;
  rarity: ERarity;
  requirement: string;
  tags: string[];
  tier: number;
  upgradeSlots: number;
  vendorPrice: number;
}