// webapp/src/api/model/shield.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { Dice } from './dice';
import type { ShieldEffectsItem } from './shieldEffectsItem';
import type { Material } from './material';
import type { ERarity } from './eRarity';

export interface Shield {
  armor: number;
  description: string;
  dice: Dice;
  effects: ShieldEffectsItem[];
  hit: number;
  id?: string;
  initiative: number;
  material: Material;
  maximumStackSize: number;
  minimumStackSize: number;
  name: string;
  note: string;
  protection: number;
  rarity: ERarity;
  requirement: string;
  tags: string[];
  tier: number;
  upgradeSlots: number;
  vendorPrice: number;
  weight: number;
}